requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
openai
duckduckgo-search
playwright
